Use in setters too
